Use agent to Google Search Linkedin Profile URL
This code uses the SerpAPI as a LangChain Tool that is used by an Agent
inside of a Chain. The goal is to obtain the Linkedin Profile URL
passing a string such as the name of the person to the LangChain agent.

The agent will use ChatGPT in order to interpret the results coming from
Google, reason if the result provides the profile URL, and iterate until
it does (or it stops).

After obtaining the profile URL, Nubela's ProxyCurl is used to scrape
the Linkedin profile. Then a simple chain (no agents) is used to
interact with ChatGPT and summarize some of the profile information.